Update humidifier-flow
Converted msg.payload objects to variables in the function. Created 2 new variables for the Humidity on and off percent. Updated the function to output the new values to a third function output everytime the function is ran. Using the third output to report these on/off values back to Home Assistant into an Input_number helper.